refactor: extract independent services from bridge.go (Phase 1)
Agent-Task: Bridge God Object refactoring Phase 1
Agent-Decision: Extract SSL/Template/Task/Notification to separate files while keeping methods on Bridge struct to satisfy mcp.Deployer interface
Agent-Limitation: Methods still use (b *Bridge) receiver; Phase 4 will introduce proper service interfaces
